// ui/src/interfaces/Hawk.ts
export interface Hawk {
    behaviorDescription: string
    colorDescription: string
    gender: string
    habitatDescription: string
    id: number
    lengthBegin: number
    lengthEnd: number
    name: string
    pictureUrl: string
    size: string
    weightBegin: number
    weightEnd: number
    wingspanBegin: number
    wingspanEnd: number
}